algo-trade/requirements.txt
# file: requirements.txt
# Core async runtime
aiohttp==3.9.5

# Data processing
numpy==2.2.4

# Configuration
PyYAML==6.0.1

# Testing
pytest==8.2.2
pytest-asyncio==0.23.7
pytest-cov==5.0.0

# Code quality
mypy==1.10.0
flake8==7.0.0
black==24.4.2
isort==5.13.2

# Optional: Webull broker (comment out if not using)
# webull==0.2.16

# Optional: Robinhood broker (comment out if not using)
# robin-stocks==3.0.6

# Type stubs
types-PyYAML==6.0.12.20240311
